LayoutTests/imported/w3c/web-platform-tests/css/cssom/StyleSheetList-constructable-with-style-recalc-expected.txt


FAIL document.styleSheets does not include adopted style sheets assert_equals: expected 1 but got 2
